requirements.txt
beautifulsoup4==4.9.3
bs4==0.0.1
Django==2.2.10
djangorestframework==3.12.4
lxml==4.6.3
psycopg2-binary==2.8.6
requests==2.26.0
soupsieve==2.2.1
sqlparse==0.4.1
urllib3==1.26.6
